+++ packages/web-client/src/components/FolderTree.tsx
@@ -420,7 +420,7 @@ export function FolderTree({ onNewMail }: Props) {
   void lang;
 
   return (
-    <aside className="w-52 shrink-0 bg-gray-50 dark:bg-gray-900 border-r border-gray-200 dark:border-gray-700 flex flex-col h-full">
+    <aside className="w-full h-full bg-gray-50 dark:bg-gray-900 border-r border-gray-200 dark:border-gray-700 flex flex-col overflow-hidden">
       <div className="p-3">
         <button
           onClick={onNewMail}